+++ packages/doctrina-cli/src/commands/context.js
@@ -5,7 +5,7 @@ import { readdirSync } from "node:fs";
 import { exists, isDir, isFile, read, relPath, walk } from "../lib/fs-ops.js";
 import { listHeader, parseDependsOn, parseAdrScope, adrSummary } from "../lib/scan.js";
 import { getTitle, getSectionParagraph } from "../lib/doc-model.js";
-import { parseFrontmatter } from "../lib/frontmatter.js";
+import { parseFrontmatter } from "../lib/doc-model.js";
 import { flagBool, flagString, flagGivenWithoutValue } from "../lib/args.js";
 import { c } from "../lib/colors.js";
 import { GIT_STATE, historyState, changedFiles } from "../lib/git.js";